+++ Sources/StreamChat/APIClient/Endpoints/Payloads/RawJSON.swift
@@ -8,13 +8,14 @@ import Foundation
 /// Used to store and operate objects of unknown structure that's not possible to decode.
 /// https://forums.swift.org/t/new-unevaluated-type-for-decoder-to-allow-later-re-encoding-of-data-with-unknown-structure/11117
 public indirect enum RawJSON: Codable, Hashable {
-    case double(Double)
+    case number(Double)
     case string(String)
-    case integer(Int)
     case bool(Bool)
     case dictionary([String: RawJSON])
     case array([RawJSON])
     case `nil`
+
+    static let double = number
 
     public init(from decoder: Decoder) throws {
         let singleValueContainer = try decoder.singleValueContainer()
@@ -24,11 +25,8 @@ public indirect enum RawJSON: Codable, Hashable {
         } else if let value = try? singleValueContainer.decode(String.self) {
             self = .string(value)
             return
-        } else if let value = try? singleValueContainer.decode(Int.self) {
-            self = .integer(value)
-            return
         } else if let value = try? singleValueContainer.decode(Double.self) {
-            self = .double(value)
+            self = .number(value)
             return
         } else if let value = try? singleValueContainer.decode([String: RawJSON].self) {
             self = .dictionary(value)
@@ -49,10 +47,8 @@ public indirect enum RawJSON: Codable, Hashable {
     
     public func encode(to encoder: Encoder) throws {
         var container = encoder.singleValueContainer()
-        
         switch self {
-        case let .integer(value): try container.encode(value)
-        case let .double(value): try container.encode(value)
+        case let .number(value): try container.encode(value)
         case let .bool(value): try container.encode(value)
         case let .string(value): try container.encode(value)
         case let .array(value): try container.encode(value)